# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for matchmaking, game management, user management, and ELO rating.",
        "title": "Elo Matchmaking Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/game": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new game with matchmaking and ELO configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "summary": "Create a game",
                "parameters": [
                    {
                        "description": "Game creation payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/src_api_game.CreateGameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_andy98725_elo-service_src_models.GameResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "403": {
                        "description": "user is not allowed to create games",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "409": {
                        "description": "game name already taken",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/game/{id}": {
            "get": {
                "description": "Returns a single game by its UUID. Public — no auth required.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "summary": "Get a game by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_andy98725_elo-service_src_models.GameResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates game settings. Only the game owner can update.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "summary": "Update a game",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_andy98725_elo-service_src_models.UpdateGameParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_andy98725_elo-service_src_models.GameResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "403": {
                        "description": "caller is not the game owner",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a game. Only the game owner can delete.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "summary": "Delete a game",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "403": {
                        "description": "caller is not the game owner",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/games": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of all games. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "summary": "List all games (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 0)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 10)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "games, nextPage",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the currently authenticated user's profile",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_andy98725_elo-service_src_models.UserResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates user fields. Defaults to the authenticated user;\nadmins may target another user with `?id=\u003cuserID\u003e`.\n`can_create_game` is admin-only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target user UUID (admin only)",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/src_api_user.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_andy98725_elo-service_src_models.UserResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a user with the provided username, email, and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/src_api_user.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_andy98725_elo-service_src_models.UserResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "409": {
                        "description": "username or email already taken",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes the authenticated user's account (not yet implemented)",
                "tags": [
                    "Users"
                ],
                "summary": "Delete current user",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/game": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of games owned by the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "summary": "Get games owned by current user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "games, nextPage",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "internal": {
                    "description": "Stores the error returned by an external dependency"
                },
                "message": {}
            }
        },
        "github_com_andy98725_elo-service_src_models.GameResp": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "elo_strategy": {
                    "type": "string"
                },
                "guests_allowed": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "lobby_enabled": {
                    "type": "boolean"
                },
                "lobby_size": {
                    "type": "integer"
                },
                "matchmaking_machine_name": {
                    "type": "string"
                },
                "matchmaking_machine_ports": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "matchmaking_strategy": {
                    "type": "string"
                },
                "metadata_enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/github_com_andy98725_elo-service_src_models.UserResp"
                },
                "public_match_logs": {
                    "type": "boolean"
                },
                "public_results": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_andy98725_elo-service_src_models.UpdateGameParams": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "elo_strategy": {
                    "type": "string"
                },
                "guests_allowed": {
                    "type": "boolean"
                },
                "lobby_enabled": {
                    "type": "boolean"
                },
                "lobby_size": {
                    "type": "integer"
                },
                "matchmaking_machine_name": {
                    "type": "string"
                },
                "matchmaking_machine_ports": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "matchmaking_strategy": {
                    "type": "string"
                },
                "metadata_enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "public_match_logs": {
                    "type": "boolean"
                },
                "public_results": {
                    "type": "boolean"
                }
            }
        },
        "github_com_andy98725_elo-service_src_models.UserResp": {
            "type": "object",
            "properties": {
                "can_create_game": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "src_api_game.CreateGameRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "elo_strategy": {
                    "type": "string"
                },
                "guests_allowed": {
                    "type": "boolean"
                },
                "lobby_enabled": {
                    "type": "boolean"
                },
                "lobby_size": {
                    "type": "integer"
                },
                "matchmaking_machine_name": {
                    "type": "string"
                },
                "matchmaking_machine_ports": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "matchmaking_strategy": {
                    "type": "string"
                },
                "metadata_enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "public_match_logs": {
                    "type": "boolean"
                },
                "public_results": {
                    "type": "boolean"
                }
            }
        },
...
        "src_api_user.RegisterRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "src_api_user.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "can_create_game": {
                    "description": "CanCreateGame is admin-only. Pointer so we can distinguish \"field\nomitted\" from \"explicit false\" — non-admins setting it (either value)\ngets a 403.",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Prefix with \"Bearer \".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}